chore(lib-auth): move to lib folder and use typescript aliases
